+++ src/shapes/Text/constants.ts
@@ -1,4 +1,4 @@
-import { LEFT } from '../../constants';
+import { LEFT, reNewline } from '../../constants';
 import type { TClassProperties } from '../../typedefs';
 import type { Text } from './Text';
 
@@ -63,7 +63,7 @@ export const styleProperties: Readonly<StylePropertiesType[]> = [
 // regexes, list of properties that are not suppose to change by instances, magic consts.
 // this will be a separated effort
 export const textDefaultValues: Partial<TClassProperties<Text>> = {
-  _reNewline: /\r?\n/,
+  _reNewline: reNewline,
   _reSpacesAndTabs: /[ \t\r]/g,
   _reSpaceAndTab: /[ \t\r]/,
   _reWords: /\S+/g,