+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [next]
 
+- fix(Text): styles line break [#8973](https://github.com/fabricjs/fabric.js/pull/8973)
 - fix(): regression to itext focusing from #8939 [#8970](https://github.com/fabricjs/fabric.js/pull/8970)
 - ci(): warn build errors in dev mode [#8971](https://github.com/fabricjs/fabric.js/pull/8971)
 - ci(): invoke tests after changelog action [#8974](https://github.com/fabricjs/fabric.js/pull/8974)